+++ tests/testthat/test-dataframe.R
@@ -896,3 +896,11 @@ test_that("pivot args works", {
     )
   )
 })
+
+test_that("rename", {
+  a = pl$DataFrame(mtcars)$rename(c("mpg", "hp"), c("miles_per_gallon", "horsepower"))$columns
+  expect_false("hp" %in% a)
+  expect_false("mpg" %in% a)
+  expect_true("miles_per_gallon" %in% a)
+  expect_true("horsepower" %in% a)
+})