+++ NEWS.md
@@ -1,3 +1,8 @@
+# polars (development version)
+## What's changed
+
+- New function `rename()` for `LazyFrame` and `DataFrame` (#239)
+
 # polars 0.6.1
 ## What's changed
  - use `pl$set_polars_options(debug_polars = TRUE)` to profile/debug method-calls of a polars query (#193)